feat: create domain modal has been seperated
